requirements.txt
pandas
numpy
pillow
pyyaml

torch
torchvision
timm

fastapi
uvicorn[standard]
python-multipart
